refactor: extract active deposit to own component
